Replace UnauthorizedException with AuthorizationException
In the FileController, the UnauthorizedException was replaced with the standard Laravel AuthorizationException for enhanced compatibility and to follow Laravel's standard design pattern. UnauthorizedException requires additional package and it was removed to minimize dependencies with 3rd party packages. Also, the function execution for external request in getContentsFromURL function has been updated to use a more descriptive method call of simpleHttp.