Refactor File Model and remove soft deletes
Removed the use of soft deletes package in the File Model. This was done because we no longer require the functionality of restoring the files once deleted. Also, replaced the user method and property with owner using ModelOwnedTrait for better understanding and readability in accordance to the new package starter-kit. Adjusted migrations and related files accordingly to reflect the removal of soft deletes and user. Also modified FileController to remove unnecessary events and authorization checks, adopting the same owner strategy.